// src/lib/api/models/RoleDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { PermissionDto } from './PermissionDto';
export type RoleDto = {
    id?: number;
    title?: string | null;
    description?: string | null;
    permissions?: Array<PermissionDto> | null;
};